% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dynamics.R
\name{compute_dynamics}
\alias{compute_dynamics}
\title{Compute emotion dynamics for two-dimensional data}
\usage{
compute_dynamics(
  data,
  x,
  y,
  id = NULL,
  time = NULL,
  level = 0.68,
  disp_length_min = 3
)
}
\arguments{
\item{data}{A data frame or tibble containing the data.}

\item{x}{A numeric variable contained in the data object}

\item{y}{A second numeric variable contained in the data object}

\item{id}{If there are multiple groups (e.g., data from many participants) this will perform the operation by group}

\item{time}{A variable in data identifying temporal ordering. If NULL data is assumed to be ordered.}

\item{level}{alpha level at which to calculate confidence ellipse for home base (defaults to .68 or +- 1 SD).}

\item{disp_length_min}{Minimal number of time points that a displacement can have. Must be between 1 and length of data,
but we recommend at least 3.}
}
\value{
tibble
}
\description{
compute_dynamics() takes a data frame of time-ordered (or user-specified time-ordered) data. In the UED processing pipeline this function
is called after text pre-processing. This means that this function can be called on any kind of numeric data.
}
