% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ued.R
\name{ued}
\alias{ued}
\title{Utterance Emotion Dynamics}
\usage{
ued(
  data,
  text,
  lexicon = "vad",
  id = NULL,
  time = NULL,
  stop_words = NULL,
  roll_avg = 10,
  level = 0.68,
  disp_length_min = 3,
  summarise = FALSE,
  min_count = 0
)
}
\arguments{
\item{data}{A data frame or tibble containing the data.}

\item{text}{Input text column.}

\item{lexicon}{Emotion lexicon. Currently VAD lexicon is only option.}

\item{id}{Optional: id column for grouping.}

\item{time}{Optional: Column indicating the temporal ordering of the text.
If the data are pre-arranged in temporal order this can be ignored.}

\item{stop_words}{Optional: character string or data frame containing stop words.
If entering a data frame, function will use the first character column it sees.}

\item{roll_avg}{Number of words to use in the rolling average.}

\item{level}{What alpha level should be used to compute the home base? Defaults to 0.68 corresponding to +- 1 SD.}

\item{disp_length_min}{Minimum number of words for displacements.}

\item{summarise}{If TRUE, output will be summarised such that each row corresponds to an id with each column representing
the mean for the particular metric. If FALSE, output will contain a row for each word.}

\item{min_count}{should words below this count be removed?}

\item{va_limits}{A tuple of scores of the VA lexicon to trim. Will exclude scores between the two values (inclusive).
This is useful for eliminating "neutral" VA terms.}
}
\value{
data.frame or tibble
}
\description{
Takes a data frame of utterances (e.g., tweets, sentences, dialogue, words, etc.) and computes Utterance Emotion Dynamics.
This function can be used to compute UED metrics for the whole dataset or groups in the dataset.
}
